requirements.txt
google-api-python-client
yt-dlp
faster-whisper
openai
google-genai
langchain
langchain-google-genai
langchain-openai
tinydb
isodate
streamlit>=1.40.0
python-dotenv
static-ffmpeg
scikit-learn
numpy
